+++ tests/tests_test_workflow/test_dependency_installer_opensearch.py
@@ -1,11 +1,11 @@
 import os
-import threading
 import unittest
 from unittest.mock import call, patch
 from urllib.error import HTTPError
 
 from manifests.build_manifest import BuildManifest
 from manifests.bundle_manifest import BundleManifest
+from system.thread_safe_counter import ThreadSafeCounter
 from test_workflow.dependency_installer_opensearch import DependencyInstallerOpenSearch
 
 
@@ -15,24 +15,11 @@ class DependencyInstallerOpenSearchTests(unittest.TestCase):
     DIST_MANIFEST_LOCAL = os.path.join(DATA, "local", "dist", "opensearch", "manifest.yml")
     DIST_MANIFEST_REMOTE = os.path.join(DATA, "remote", "dist", "opensearch", "manifest.yml")
 
-    class ThreadSafeCounter:
-        def __init__(self):
-            self.lock = threading.Lock()
-            self.__call_count__ = 0
-
-        def thread_safe_count(self, *args, **kwargs):
-            with self.lock:
-                self.__call_count__ += 1
-
-        @property
-        def call_count(self):
-            return self.__call_count__
-
     @patch("os.makedirs")
     @patch("shutil.copyfile")
     @patch("urllib.request.urlretrieve")
     def test_install_maven_dependencies_local(self, mock_request, mock_copyfile, mock_makedirs):
-        counter = self.ThreadSafeCounter()
+        counter = ThreadSafeCounter()
         mock_copyfile.side_effect = counter.thread_safe_count
 
         dependency_installer = DependencyInstallerOpenSearch(
@@ -64,7 +51,7 @@ class DependencyInstallerOpenSearchTests(unittest.TestCase):
     @patch("shutil.copyfile")
     @patch("urllib.request.urlretrieve")
     def test_install_maven_dependencies_remote(self, mock_request, mock_copyfile, mock_makedirs):
-        counter = self.ThreadSafeCounter()
+        counter = ThreadSafeCounter()
         mock_request.side_effect = counter.thread_safe_count
         dependency_installer = DependencyInstallerOpenSearch(
             "https://ci.opensearch.org/x/y",
